luarocks make: move rockspec picking into an util function
